New Preset and slightly different approach...
There are now 3 Presets that can be switched with the arrows on the
right.
Navigation is done by holding the stop button down and then track bank,
cursor track or device can be moved with RWD and FFW. Stop & Loop
together allows to switch between modes. Changing preset and wiggling a
knob will also switch to that mode.